requirements.txt
PyVCF
biopython
vcf-parser
vcfpy
